클라우드센터 가상머신 live 상태에서 디스크  스냅샷 백업 가능하도록 기능 개선
기존에 CCVM이 정지 상태인 경우에만 정지 가능하였는데, "Strated", "Stopped" 상태 모두 디스크 스냅샷 백업 가능하도록 기능 개선

+++ python/ccvm_snap/ccvm_snap_action.py
@@ -14,6 +14,8 @@ import os
 import json
 import datetime
 import subprocess
+import sh
+import socket
 from subprocess import check_output
 
 from ablestack import *
@@ -40,7 +42,7 @@ def createArgumentParser():
 
     # 인자 추가: https://docs.python.org/ko/3/library/argparse.html#the-add-argument-method
 
-    parser.add_argument('action', choices=['list', 'rollback', 'backup'], help='choose one of the actions')
+    parser.add_argument('action', choices=['list', 'rollback', 'backup', 'test'], help='choose one of the actions')
     parser.add_argument('--snap-name', metavar='snapshot name', type=str, help='The name of the CCVM Snapshot')
 
     # output 민감도 추가(v갯수에 따라 output및 log가 많아짐):
@@ -88,26 +90,47 @@ def backupCcvmSnap(args):
     # 실행중인 ccvm 백업할 경우 추가작업 : ccvm이 실행중인 호스트를 확인하여 virsh 명령으로 suspend 후 스냅샷 생성하고 완료후 다시 resume 하는 로직 필요
     try:
         now = datetime.datetime.now().strftime('%Y-%m-%d-%H:%M:%S')
+        pcs_status = json.loads(sh.python3(pluginpath + "/python/pcs/main.py","status", "--resource", "cloudcenter_res").stdout.decode())
+        pcs_started_host_name = pcs_status['val']['started']
+        pcs_status = pcs_status['val']['role']
+        ccvm_name = 'ccvm'
+        ccvm_image_name = 'ccvm'
+        pool_name = 'rbd'
 
-        # ccvm 스냅 생성
-        result = os.system("rbd snap create "+pool_name+"/"+ccvm_image_name+"@"+now)
+        if pcs_started_host_name == 'None' or pcs_status == 'Started' or pcs_status == 'Stopped':
+            
+            if pcs_status == 'Started':
+                # ccvm suspend
+                os.system("ssh root@"+pcs_started_host_name+" \"virsh suspend "+ccvm_name+" > /dev/null\"")
 
-        # ccvm 스냅이 10개 이상이면 마지막 스냅 삭제
-        output = check_output(["rbd snap list "+ccvm_image_name+" --format json"], universal_newlines=True, shell=True, env=env)
-        output_json = json.loads(output)
+                # ccvm 스냅 생성
+                result = os.system("rbd snap create "+pool_name+"/"+ccvm_image_name+"@"+now)
+                
+                # ccvm resume
+                os.system("ssh root@"+pcs_started_host_name+" \"virsh resume "+ccvm_name+" > /dev/null\"")
+            else:
+                # ccvm 스냅 생성
+                result = os.system("rbd snap create "+pool_name+"/"+ccvm_image_name+"@"+now)
 
-        ccvm_snap_cnt = len(output_json)
-        ccvm_snap_limit = 10
+            # ccvm 스냅이 10개 이상이면 마지막 스냅 삭제
+            output = check_output(["rbd snap list "+ccvm_image_name+" --format json"], universal_newlines=True, shell=True, env=env)
+            output_json = json.loads(output)
 
-        for ccvm_snap_info in output_json:
-            if ccvm_snap_cnt > ccvm_snap_limit:
-                ccvm_snap_cnt = ccvm_snap_cnt - 1
-                os.system("rbd snap rm "+pool_name+"/"+ccvm_image_name+"@"+ccvm_snap_info["name"])
+            ccvm_snap_cnt = len(output_json)
+            ccvm_snap_limit = 10
 
-        if result == 0:
-            return createReturn(code=200, val="CCVM Snapshot Backup Create Success")
-        else:
-            return createReturn(code=500, val="CCVM Snapshot Backup Create Fail. Check if ccvm is Stopped")
+            for ccvm_snap_info in output_json:
+                if ccvm_snap_cnt > ccvm_snap_limit:
+                    ccvm_snap_cnt = ccvm_snap_cnt - 1
+                    os.system("rbd snap rm "+pool_name+"/"+ccvm_image_name+"@"+ccvm_snap_info["name"])
+
+            if result == 0:
+                return createReturn(code=200, val="CCVM Snapshot Backup Create Success")
+            else:
+                return createReturn(code=500, val="CCVM Snapshot Backup Create Fail. Snapshot backup is possible only in the stop or start state1")
+
+        else :
+            return createReturn(code=500, val="CCVM Snapshot Backup Create Fail. Snapshot backup is possible only in the stop or start state")
 
     except Exception as e:
         # 결과값 리턴
@@ -132,4 +155,3 @@ if __name__ == '__main__':
         print(rollbackCcvmSnap(args))
     elif (args.action) == 'backup':
         print(backupCcvmSnap(args))
-
